iterate stage1 utf8 code
The optimized x86_64 code only runs on Linux, because it is
SSE4.1, which X86_64 only guarantees SSE2, and so needs CPU detection
support, and on Linux that support available is out-of-the-box with
the IFUNC ELF extension. This is not dissimmilar to how we do vdso
clock_gettime() in std/os.zig with a resolver function, but it has
a dedicated ELF section for the jump table, that gets resolved
at program startup.

Also has a arm64 neon implementation.

Invalid UTF-8 does not (currently) report the invalid character.

Also validats the no control character rule. (including hard tabs)