std.unicode rework, unified error codes
Use Zig's error unions:

There is still a general error.InvalidUtf8 that gets thrown instead
of generalized codes by Utf8View.init(). Changing this would
change the error all many of the os.file functions that support
Windows. 5 errors is a bit much for that interface, when usually
the consumer doesn't care which. To make this more managable I
proposed to allow switch to be used on an error union, which
Zig would compile into a switch on all members of that union. #2228